services/adminlogin/requirements.txt
flask
grpcio
mysql-connector-python
python-dotenv
protobuf
bcrypt
psycopg2-binary
Flask-Cors